refactor: utils structure has been updated
